// src/lib/types.ts
export type SocialMetric = {
  id: string;
  platform: string;
  businessUnit: string;
  date: string;
  metric: string;
};  

export type WebsiteMetric = {
  id: string;
  businessUnit: string;
  date: string;
  metric: string;
};

export type NewsletterMetric = {
  id: string;
  businessUnit: string;
  date: string;
  metric: string;
};